Rename ikfast demo and its files
