+++ IKFastDemo/Properties/AssemblyInfo.cs
