+++ RobotComponents.sln
@@ -19,7 +19,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RobotComponents", "RobotCom
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RobotComponents.ABB.Presets", "RobotComponents.ABB.Presets\RobotComponents.ABB.Presets.csproj", "{15938314-61C9-44E9-B0E9-6E5CB15825F6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IkgenDemo", "IkgenDemo\IkgenDemo.csproj", "{E279A641-7002-47FF-A0EE-694CAB972A5C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IKFastDemo", "IKFastDemo\IKFastDemo.csproj", "{E279A641-7002-47FF-A0EE-694CAB972A5C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IKFastTest", "IKFastTest\IKFastTest.csproj", "{32FDB79E-9D2B-4425-89CA-576163380866}"
 EndProject
